fix t_color class{double -> int}

+++ include/color.h
@@ -10,9 +10,9 @@
 
 typedef struct s_color
 {
-	double	r;
-	double	g;
-	double	b;
+	int	r;
+	int	g;
+	int	b;
 }	t_color;
 
 t_error	set_color(t_color *color, int r, int g, int b);
